Add FWHM to velocity.gaussians montecarlo
Adds FWHM calculation to velocity.gaussians using montecarlo for uncertainty. Template can be used to implement this in other gaussian methods. Modifies gaussians.py iterator created by the @BaseToolGaussians decorater call method to return also the deepcopy and uncertainty of the stddevs, which are needed to calculate FWHM for a gaussian. It's worth checking whether this additional return taking the shape of the return from 3 to 5 changes or breaks anything. These additional returns are added to the end and have been verified to work with headless and GUU velocity.gaussians montecarlo.

Currently, the calculation is only done for ASCII mode and not for JSON.